plthook: Setup module id for BIND-NOW binaries
The PLT resolver (and our plt_hooker) needs module id to identify
a function among the libraries.  But BIND-NOW doesn't need this
machinary so module id (at GOT[1]) was set to 0.  But uftrace still
needs to distinguish the modules so use a pointer to plthook_data as
module id instead.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>