symbol: Fix SYMTAB_FL_ADJ_OFFSET usage in plthook
The SYMTAB_FL_ADJ_OFFSET is to adjust offset to calculate symbol address
from an ELF file.  It adds an offset to symbol value to calcuate the
real (memory) address like below.

The offset is a difference between loaded address (as can be shown in
/proc/self/maps) and virtual address in ELF file (phdr.p_vaddr).  It
seems dl_iterate_phdr() gives us the offset directly, but if we only
have the load address, it needs to be adjusted when loading symbols.

For a normal executable file, the addresses are looked like below:

  load addr: 0x400000 (= /proc/self/maps)
  virt addr: 0x400000 (= phdr.p_vaddr)
  sym  addr: 0x401234 (= sym.st_value)

  real addr: 0x401234  (= sym addr + (load addr - virt addr))
                                      ~~~~~~~~~~~~~~~~~~~~~~~ = offset

  file addr: 0x002134  (= real addr - load addr = sym addr - virt addr)
  user addr: 0x401234  (= real addr = file addr + load addr)

For a shared library or a PIE binary:

  load addr: 0x100000 (= in /proc/maps)
  virt addr: 0x000000 (= phdr.p_vaddr)
  sym  addr: 0x003458 (= sym.st_value)

  real addr: 0x103458  (= sym addr + (load addr - virt addr))
                                      ~~~~~~~~~~~~~~~~~~~~~~~ = offset
  file addr: 0x003458  (= real addr - load addr)
  user addr: 0x103458  (= file addr + load addr)

It should work for prelinked libraries but I didn't check.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>