symbol: Load and save all module symbols
With --nest-libcall uftrace can trace (library) functions from any
modules.  To make user easier to setup filters or triggers, it needs to
apply them to all available modules.  But it's tricky to determine which
one is available.  Since symbol data is not big (compared to trace
data), it'd rather save all symbols unconditionally.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>