symbol: Save correct (relative) symbol address
The symbol files intended to save relative addresses so that they can be
recalculated to the real address using map info in a session.  But it
was found that it failed to do so.

The SYMTAB_FL_ADJ_OFFSET flag was to adjust current load address with
ELF load address to calcuate real symbol address with the
offset (especially for main executables and prelink-ed libraries).

It also failed to save relative addresses for dlopen-ed libraries.  So
it makes the dlopen test case failed - fix it.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>