misc: Add args/retval format generator from prototype
This adds a list prototypes of well-known library functions in
"prototypes.h" and its parser that generates uftrace format argspec.

The usage is as follows:
  $ ./misc/gen-autoargs.py -i ./misc/prototypes.h

Then it generates "autoargs.h" that can be directly included by uftrace
for --auto-args option support.

In the below example, "simple_prototypes.h" has only 5 functions.  It
can be passed to "gen-autoargs.py" and the output is as follows:

  $ cat simple_prototypes.h
  void *malloc(size_t size);
  void free(void* ptr);
  const char* strcpy(char* dest, const char* src);
  int open(const char* pathname, int flags);
  int pthread_mutex_lock(pthread_mutex_t *mutex);

  $ ./gen-autoargs.py simple_prototypes.h
    GEN      autoargs.h

  $ cat autoargs.h
      ...
  static char *auto_args_list =
          "malloc@arg1/u;"
          "free@arg1/x;"
          "strcpy@arg1/s,arg2/s;"
          "open@arg1/s,arg2;"
          "pthread_mutex_lock@arg1/x;"
  ;

  static char *auto_retvals_list =
          "malloc@retval/x;"
          "strcpy@retval/s;"
          "open@retval;"
          "pthread_mutex_lock@retval;"
  ;

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>